Add device name editor, e-paper touch responsiveness, and 64-bit cleanup
Features:
- Settings: new Device Name section persisted via RnsService::userInfo
  with safe ArduinoJson String storage and re-announce on change
- Companion server: --reset flag for interactive identity/trust/config wipe
- Companion server TUI: fire service events on trust offers, message
  delivery, and delivery failures so they surface in the UI
- E-paper: register GxEPD2 busy callback so LVGL touch input is processed
  during the ~700ms refresh, allowing scroll gestures to track the finger

Fixes:
- Replace 32-bit pointer casts in Launcher/UChat with (uintptr_t) so the
  64-bit emulator builds without -fpermissive
- Drop -fpermissive from emulator and test envs; add -std=gnu++17 and
  LV_DISP_DEF_REFR_PERIOD=50 for the hardware build
- RnsService: report DeserializationError and log why a payload was not
  recognized as a service message
- test_lxmf_workflow: store packet.raw() in a local Bytes to avoid using
  a temporary across .find()

Docs/tests:
- CLAUDE.md: document propagation service, action queue, coding standards
  (ArduinoJson string storage, temporaries, null safety, cross-thread
  safety), expanded test inventory, and a Lessons Learned section
- New test_device_name suite covering String safety and announce format
- .gitignore: ignore .claude and .serena

+++ src/apps/Launcher.cpp
@@ -30,7 +30,7 @@
 
 static void menu_btn_event_cb(lv_event_t *e)
 {
-    uint32_t id = (uint32_t)e->user_data;
+    uint32_t id = (uint32_t)(uintptr_t)e->user_data;
     Serial.print("APP CLICKED=");
     Serial.println(id);
     // lv_label_set_text(label, tgr->name);
@@ -72,6 +72,6 @@ void Launcher::menu_btn_create(lv_obj_t *parent, AppInfo& appInfo, int x, int y)
     lv_obj_set_style_border_width(label, 0, LV_PART_MAIN | LV_STATE_DEFAULT);
     //launch on click
     uint32_t id_hax = appInfo.id;
-    lv_obj_add_event_cb(btn, menu_btn_event_cb, LV_EVENT_CLICKED, (void *)(id_hax));
+    lv_obj_add_event_cb(btn, menu_btn_event_cb, LV_EVENT_CLICKED, (void *)(uintptr_t)(id_hax));
     
 }
